+++ nodes/Ragmetrics/index.ts
@@ -1,4 +1,4 @@
-export * from './Jaasai.node';
+export * from './Ragmetrics.node';
 // If you have helpers or interfaces, you can also export them:
 // export * from './helpers';
 // export * from './interfaces';